comments, cleanup, some slight refactoring
