+++ src/main/java/org/cancogenvirusseq/muse/api/ApiDefinition.java
@@ -122,8 +122,8 @@ public interface ApiDefinition {
       consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
       produces = MediaType.APPLICATION_JSON_VALUE,
       method = RequestMethod.POST)
-  @HasReadWriteAccess
   @Transactional
+  @HasReadWriteAccess
   Mono<SubmissionCreateResponse> submit(
       // this isn't working correctly in the swagger-ui, know issue:
       // https://github.com/springfox/springfox/issues/3464
